+++ Sources/StreamChat/WebSocketClient/Events/EventPayload.swift
@@ -35,6 +35,7 @@ class EventPayload: Decodable {
         case lastDeliveredAt = "last_delivered_at"
         case lastDeliveredMessageId = "last_delivered_message_id"
         case unreadMessagesCount = "unread_messages"
+        case groupedUnreadChannels = "grouped_unread_channels"
         case shadow
         case thread
         case vote = "poll_vote"
@@ -74,6 +75,7 @@ class EventPayload: Decodable {
     let lastDeliveredAt: Date?
     let lastDeliveredMessageId: MessageId?
     let unreadMessagesCount: Int?
+    let groupedUnreadChannels: GroupedUnreadChannels?
     var poll: PollPayload?
     var vote: PollVotePayload?
 
@@ -102,6 +104,7 @@ class EventPayload: Decodable {
         reaction: MessageReactionPayload? = nil,
         watcherCount: Int? = nil,
         unreadCount: UnreadCountPayload? = nil,
+        groupedUnreadChannels: GroupedUnreadChannels? = nil,
         createdAt: Date? = nil,
         isChannelHistoryCleared: Bool? = nil,
         banReason: String? = nil,
@@ -151,6 +154,7 @@ class EventPayload: Decodable {
         self.lastReadAt = lastReadAt
         self.lastReadMessageId = lastReadMessageId
         self.unreadMessagesCount = unreadMessagesCount
+        self.groupedUnreadChannels = groupedUnreadChannels
         self.threadPartial = threadPartial
         self.threadDetails = threadDetails
         self.poll = poll
@@ -194,6 +198,7 @@ class EventPayload: Decodable {
         lastReadAt = try container.decodeIfPresent(Date.self, forKey: .lastReadAt)
         lastReadMessageId = try container.decodeIfPresent(MessageId.self, forKey: .lastReadMessageId)
         unreadMessagesCount = try container.decodeIfPresent(Int.self, forKey: .unreadMessagesCount)
+        groupedUnreadChannels = try container.decodeIfPresent(GroupedUnreadChannels.self, forKey: .groupedUnreadChannels)
         threadDetails = container.decodeAsResultIfPresent(ThreadDetailsPayload.self, forKey: .thread)
         threadPartial = container.decodeAsResultIfPresent(ThreadPartialPayload.self, forKey: .thread)
         vote = try container.decodeIfPresent(PollVotePayload.self, forKey: .vote)
@@ -239,6 +244,7 @@ private extension PartialKeyPath where Root == EventPayload {
         case \EventPayload.reaction: return "reaction"
         case \EventPayload.watcherCount: return "watcherCount"
         case \EventPayload.unreadCount: return "unreadCount"
+        case \EventPayload.groupedUnreadChannels: return "groupedUnreadChannels"
         case \EventPayload.createdAt: return "createdAt"
         case \EventPayload.isChannelHistoryCleared: return "isChannelHistoryCleared"
         case \EventPayload.banReason: return "banReason"