+++ Tests/StreamChatTests/ChatClient_Tests.swift
@@ -266,7 +266,6 @@ final class ChatClient_Tests: XCTestCase {
 
     func test_groupedQueryChannels_callsAPIClientAndReturnsGroupedChannels() {
         let client = ChatClient.mock(config: inMemoryStorageConfig)
-        client.setToken(token: try! .init(rawValue: "eyJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoiZHVtbXlDdXJyZW50VXNlciJ9.signature"))
         let firstCid = ChannelId.unique
         let secondCid = ChannelId.unique
         let thirdCid = ChannelId.unique
@@ -318,8 +317,7 @@ final class ChatClient_Tests: XCTestCase {
         XCTAssertEqual(receivedGroupedChannels?.channels["all"]?.map(\.cid), [firstCid])
         XCTAssertEqual(receivedGroupedChannels?.channels["new"]?.map(\.cid), [secondCid])
         XCTAssertEqual(receivedGroupedChannels?.channels["current"]?.map(\.cid), [thirdCid])
-        XCTAssertEqual(receivedGroupedChannels?.channels["all"]?.first?.unreadCount.messages, 10)
-        XCTAssertEqual(receivedGroupedChannels?.unreadCounts, ["all": 10, "new": 10, "current": 10])
+        XCTAssertEqual(receivedGroupedChannels?.unreadCounts, ["all": 1, "new": 2, "current": 4])
         XCTAssertEqual(receivedGroupedChannels?.unreadChannels, ["all": 1, "new": 1, "current": 2])
     }